% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairdata.R
\name{pairdata}
\alias{pairdata}
\title{Generate paired dataset from orginal dataset}
\usage{
pairdata(data1)
}
\arguments{
\item{data1}{the orginal unpaired data}
}
\description{
Generate the paired data used for logisitic regression
}
\examples{
n <- 50
p <- 8
beta <- c(3,0,1.5,0,2,rep(0,p-5))
xm <- matrix(rnorm(n*p),ncol = p, nrow = n)
y <- xm \%*\% beta + rnorm(n)
colnames(xm) <- paste0("Var_",1:p)
pair01 <- pairdata(cbind(y,xm))
nrow(pair01) == choose(n,2)

}
