% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairdata.R
\name{pairdata}
\alias{pairdata}
\title{Generate paired dataset from orginal dataset}
\usage{
pairdata(data1)
}
\arguments{
\item{data1}{the orginal unpaired data}
}
\description{
Generate the paired data used for logisitic regression
}
